extract dhis logic
